ai-chat: support dragging conversations to tabs
